Split up urls into apps
